Add support for site with translations
This kind of hackily supports translations for footer_links/tagline.
An example of how that might look when you've got english and russian
translations:

[extra]
[extra.footer_tagline]
en = "English tagline"
ru = "Russian tagline"

[extra.footer_links]
en = [{url = "$BASE_URL/about", name = "English about"}]
ru = [{url = "$BASE_URL/about", name = "Russian about"}]

If this isn't needed the old way that the readme describes still works,
then the same values are use independently of the language on the rest
of the page. The $BASE_URL in the links above will change depending on
the current language.

The 'title' and 'description' items should be placed in [translations]
and [languages.XX] as described by the Zola guide.